feat: update test suit for currency entity
